% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy-data.R
\docType{data}
\name{clusterBAData}
\alias{clusterBAData}
\title{a cluster sample data object}
\format{an object of class dataframe
\itemize{
  \item{clusterID}{cluster}
  \item{bapa}{basal area per acre - attribute of interest}
  \item{isUsed}{TRUE/FALSE}
}}
\usage{
clusterBAData
}
\description{
A more complex data example of a cluster sample
}
