% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy-data.R
\docType{data}
\name{dataPlot}
\alias{dataPlot}
\title{a cluster sample data object}
\format{an object of class dataframe
\itemize{
  \item{clusterID}{cluster}
  \item{volume}{sampled volume - attribute of interest}
  \item{isUsed}{TRUE/FALSE}
}}
\usage{
dataPlot
}
\description{
A simple data example of a cluster sample, drawn from
Avery and Burkhart Forest Measurements text.
}
