% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_cluster.R
\name{summarize_all_cluster}
\alias{summarize_all_cluster}
\title{Summarize all cluster sample}
\usage{
summarize_all_cluster(data, attribute = NA, plot = TRUE, plotTot = NA,
  desiredConfidence = 0.95, bernoulli = F)
}
\arguments{
\item{data}{dataframe containing observations of variable of
interest for either cluster-level or plot-level data.}

\item{attribute}{character name of attribute to be summarized.}

\item{plot}{logical true if parameter data is plot-level, false if
parameter data is cluster-level. Default is True.}

\item{plotTot}{numeric population size. Equivalent to the 
total number of possible plots in the population.}

\item{desiredConfidence}{numeric desired confidence level (e.g. 0.9).}

\item{bernoulli}{boolean TRUE if data fitting the Bernoulli 
distribution is used.}
}
\value{
dataframe of stand-level statistics including
standard error and confidence interval limits.
}
\description{
Summarizes population-level statistics for
cluster sample data.
}
\examples{
\dontrun{
dataPlot <- data.frame(clusterID = c(1, 1, 1, 1, 1, 2, 2, 3, 4, 4, 4, 4, 4, 4, 5, 5, 5, 5, 5),
                       attr = c(1000, 1250, 950, 900, 1005, 1000, 1250, 950, 900, 1005, 1000,
                                1250, 950, 900, 1005, 1000, 1250, 950, 900),
                       isUsed = c(T, T, T, T, T, T, T, T, T, T, T, T, T, T, F, F, F, F, F))
plot = TRUE
attribute = 'attr'
}
}
\author{
Karin Wolken
}
