% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_srs.R
\name{summarize_all_srs}
\alias{summarize_all_srs}
\title{Summarize all simple random sample}
\usage{
summarize_all_srs(data, attribute = "attr", type = "vector", popSize = NA,
  desiredConfidence = 0.95, infiniteReplacement = F, bernoulli = F)
}
\arguments{
\item{data}{data frame or vector containing observations of
variable of interest. Variable of interest must already be expanded
to the level of interest (e.g. stand-level).}

\item{attribute}{character name of attribute to be summarized.
Must be defined if data is input as a data frame.}

\item{type}{object type of data, e.g. 'dataframe' or 'vector'.}

\item{popSize}{numeric population size. Defaults to NA (unknown popSize).}

\item{desiredConfidence}{numeric desired confidence level (e.g. 0.9).}

\item{infiniteReplacement}{logical true if sample was done with replacement
or from an infinite population. False if sampled without replacement,
from a finite population. Defaults to False.}

\item{bernoulli}{logical TRUE if data fitting the Bernoulli 
distribution is used.}
}
\value{
a data frame of population mean, variance, standard error, and
high and low confidence limits.
}
\description{
Summarizes population-level statistics for
simple random sample data.
}
\examples{
\dontrun{
#trainingDataFrame
data <- data.frame(bapa = c(120, 140, 160, 110, 100, 90),
  plots = c(1, 2, 3, 4, 5, 6))
#trainingVector
data <- c(120, 140, 160, 110, 100, 90)
attribute <- 'bapa'
desiredConfidence <- 0.9
}
}
\author{
Karin Wolken
}
