% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{summarize_cluster}
\alias{summarize_cluster}
\title{Summarize cluster sample}
\usage{
summarize_cluster(cluster)
}
\arguments{
\item{cluster}{dataframe containing observations of variable of
interest.}
}
\value{
dataframe of cluster-level statistics including clusterID,
standard error, and confidence interval limits.
}
\description{
Summarizes population-level statistics for
cluster sample data. The calculations are derived from Chapter 3 in
Avery and Burkhart's (1967) Forest Measurements, Fifth Edition. The
variance terms refer to the variance of the mean.
}
\examples{
\dontrun{
cluster = data.frame(clusterID = c(1, 2, 3, 4, 5),
                     clusterElements = c(4, 2, 9, 4, 10),
                     sumAttr = c(1000, 1250, 950, 900, 1005),
                     isUsed = c(T, T, F, T, T))
}
}
\author{
Karin Wolken
}
