% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{summarize_poisson}
\alias{summarize_poisson}
\title{Summarize Poisson sample}
\usage{
summarize_poisson(data, desiredConfidence = 0.95)
}
\arguments{
\item{data}{vector containing number of instances per desired 
unit (e.g. 6 trees were alive at plot 10 -> c(6)).}

\item{desiredConfidence}{numeric desired confidence level 
(e.g. 0.9).}
}
\value{
data frame of statistics including standard error and 
confidence interval limits.
}
\description{
Summarizes population-level statistics for
Poisson sample data. The calculations are derived from open 
resources provided by Penn State Eberly College.
}
\examples{
\dontrun{

# See Forest Sampling vignette for more details

# Data can be expressed as:

data <- c(2, 3, 4, 3, 4, 5, 2, 7)

}
}
\author{
Karin Wolken
}
