% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_random.R
\name{summarize_simple_random}
\alias{summarize_simple_random}
\title{Summarize simple random sample}
\usage{
summarize_simple_random(trainingData, attribute, popSize = NA,
  desiredConfidence = 0.9, infiniteReplacement = F)
}
\arguments{
\item{trainingData}{dataframe containing observations of variable of
interest.}

\item{attribute}{character name of attribute to be summarized. Attribute
must already be expanded to the level of interest (e.g. stand-level).}

\item{popSize}{numeric population size. Defaults to NA (unknown popSize).}

\item{desiredConfidence}{numeric desired confidence level (e.g. 0.9).}

\item{infiniteReplacement}{logical true if sample was done with replacement
or from an infite population. False if sampled without replacement,
from a finite population. Assumes without replacement, from a finite
population.}
}
\value{
a dataframe of population mean, variance, standard error, and
high and low confidence limits.
}
\description{
Summarizes population-level statistics for
simple random sample data. The calculations are derived from Chapter 3 in
Avery and Burkhart's (1967) Forest Measurements, Fifth Edition. The
variance terms refer to the variance of the mean, hence the
\code{sampleSize} terms in the denominators.
}
\examples{
\dontrun{
trainingData <- data.frame(bapa = c(120, 140, 160, 110, 100, 90),
  plots = c(1, 2, 3, 4, 5, 6))
attribute = 'bapa'
desiredConfidence = 0.9
}
}
\author{
Karin Wolken
}
