% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs_discrete.R
\name{summarize_simple_random_discrete}
\alias{summarize_simple_random_discrete}
\title{Summarize simple random sample for discrete attributes}
\usage{
summarize_simple_random_discrete(data, attribute = "attr", popTot = NA,
  desiredConfidence = 0.95)
}
\arguments{
\item{data}{dataframe containing observations of variable of
interest. Attribute must be coded as either TRUE and FALSE or
1 and 0.}

\item{attribute}{character name of attribute to be summarized.}

\item{popTot}{numeric population size. Equivalent to total 
number of individuals.}

\item{desiredConfidence}{numeric desired confidence level (e.g. 0.9).}
}
\value{
dataframe of stand-level statistics. Includes standard error and
95% confidence interval.
}
\description{
Summarizes population-level statistics for
simple random sample for attribute data. The calculations are
derived from Chapter 3 in Avery and Burkhart's (1967)
Forest Measurements, Fifth Edition. The variance terms refer
to the variance of the mean.
}
\examples{
\dontrun{
data <- data.frame(alive = c(T, T, F, T, F, F),
  tree = c(1, 2, 3, 4, 5, 6))
attribute = 'alive'
popTot = 50
}
}
\author{
Karin Wolken
}
