% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified.R, R/stratified_samples.R
\name{summarize_stratified}
\alias{summarize_stratified}
\title{Summarize stratified sample}
\usage{
summarize_stratified(trainingData, attribute, stratumTab,
  desiredConfidence = 0.9, post = T)

summarize_stratified(trainingData, attribute, stratumTab,
  desiredConfidence = 0.9, post = T)
}
\arguments{
\item{trainingData}{dataframe containing observations of variable of
interest, and stratum assignment for each plot}

\item{attribute}{character name of attribute to be summarized}

\item{stratumTab}{dataframe containing acreages for each stratum}

\item{desiredConfidence}{numeric desired confidence level (e.g. 0.9)}

\item{post}{logical true if post-stratification was used}

\item{trainingData}{dataframe containing observations of variable of
interest, and stratum assignment for each plot}

\item{attribute}{character name of attribute to be summarized}

\item{stratumTab}{dataframe containing acreages for each stratum}

\item{desiredConfidence}{numeric desired confidence level (e.g. 0.9)}

\item{post}{logical true if post-stratification was used}
}
\description{
Summarizes strata- and population-level statistics for
stratified sample data. The calculations are derived from Chapter 5 in
Gregoire and Valentine's (2008) Sampling Strategies for Natural Resources
and the Environment. The variance terms refer to the variance of the mean,
hence the \code{n} terms in the deonminators.

Summarizes strata- and population-level statistics for
stratified sample data. The calculations are derived from Chapter 5 in
Gregoire and Valentine's (2008) Sampling Strategies for Natural Resources
and the Environment. The variance terms refer to the variance of the mean,
hence the \code{n} terms in the deonminators.
}
\examples{
\dontrun{
trainingData <- data.frame(bapa = c(120, 140, 160, 110, 100, 90),
  stratum = c(1, 1, 1, 2, 2, 2))
stratumTab <- data.frame(stratum = c(1, 2), acres = c(200, 50))
attribute = 'bapa'
desiredConfidence = 0.9
}
\dontrun{
trainingData <- data.frame(bapa = c(120, 140, 160, 110, 100, 90),
  stratum = c(1, 1, 1, 2, 2, 2))
stratumTab <- data.frame(stratum = c(1, 2), acres = c(200, 50))
attribute = 'bapa'
desiredConfidence = 0.9
}
}
\author{
Henry Rodman

Henry Rodman
}

