% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeP.R
\name{summarize_threeP}
\alias{summarize_threeP}
\title{Summarize 3P sample}
\usage{
summarize_threeP(data, sampSize, treeCount, BAF, cvPercent, 
                        trueNetVBAR, height = FALSE, 
                        desiredConfidence = 0.95)
}
\arguments{
\item{data}{data frame containing key tree- or plot-level information.}

\item{cvPercent}{numeric average Coefficient of Variation expressed 
as a percent (e.g. 50).}

\item{trueNetVBAR}{numeric measured net VBAR (volume to basal area 
ratio).}

\item{height}{logical TRUE if data input contains height values. 
Otherwise, FALSE assumes data contains estimateVBAR.}

\item{plot}{logical TRUE if the input data is averaged by plot, 
FALSE if data is tree-level.}

\item{desiredConfidence}{numeric desired confidence level (e.g. 0.9).}
}
\value{
data frame of statistics including standard error and 
confidence interval limits.
}
\description{
Summarizes population-level statistics for
3P sample data. The calculations are derived from Kim Iles
'A Sampler of Inventory Topics', pg 601.
}
\examples{
\dontrun{

# This function has a few options for data input:


# 1. Height is included, data is plot-level.

dataHeight <- data.frame(plotNum = c(5, 4, 3, 5, 4),
                         treeCount = c(4, 5, 6, 3, 4),
                         BAF = c(10, 10, 10, 10, 10),
                         avgTreeVBARSperPlot = c(9, 8, 7, 8, 2),
                         treeHeight = c(1, 2, 3, 4, 5))
                         
# Example: 
summarize_threeP(dataHeight, cvPercent = 50, trueNetVBAR = 10,
                 height = T, plot = T)


# 2. estimateVBAR is included (height data is not used, 
   and so is not included), data is plot-level.

dataEstimateVBAR <- data.frame(plotNum = c(1, 2, 3, 4, 5),
                               treeCount = c(4, 5, 6, 3, 4),
                               BAF = c(10, 10, 10, 10, 10),
                               avgTreeVBARSperPlot = 
                                              c(9, 8, 7, 8, 2),
                               estimateVBAR = c(1, 2, 3, 4, 5))

# Example: 
summarize_threeP(dataEstimateVBAR, cvPercent = 50, 
                 trueNetVBAR = 10, height = F, plot = T)


# 3. estimateVBAR is included, data is tree-level.

dataNotPlot <- data.frame(plotNum = c(1, 1, 2, 2, 3),
                          BAF = c(10, 10, 10, 10, 10),
                          VBAR = c(1, 2, 3, 4, 3),
                          estimateVBAR = c(1, 2, 3, 4, 5))

# Example:
summarize_threeP(dataNotPlot, cvPercent = 50, trueNetVBAR = 10, 
                 height = F, plot = F)

}
}
\author{
Karin Wolken
}
