% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage.R
\name{summarize_two_stage}
\alias{summarize_two_stage}
\title{Summarize two-stage sample}
\usage{
summarize_two_stage(data, plot = TRUE, attribute = NA,
  populationClusters = 0, populationElementsPerCluster = 0)
}
\arguments{
\item{data}{dataframe containing observations of variable of
interest for either cluster-level of plot-level data.}

\item{plot}{logical True if parameter data is plot-level, False if
parameter data is cluster-level. Default is True.}

\item{attribute}{character name of attribute to be summarized.}
}
\value{
dataframe of stand-level statistics including
standard error and confidence interval limits. All final values are
on a 'per plot' basis.
}
\description{
Summarizes population-level statistics for
two-stage sample data. The calculations are derived from Chapter 3 in
Avery and Burkhart's (1967) Forest Measurements, Fifth Edition. The
variance terms refer to the variance of the mean.
}
\examples{
\dontrun{
dataCluster <- data.frame(clusterID = c(1, 2, 3, 4, 5),
                          totClusterElements = c(5, 2, 1, 6, 5),
                          sampledElements = c(2, 2, 2, 2, 2),
                          isUsed = c(T, T, T, T, F),
                          attrSumCluster = c(1000, 1250, 950, 900, 1005))

data <- data.frame(clusterID = c(1, 1, 1, 1, 1, 2, 2, 3, 4,
                   4, 4, 4, 4, 4, 5, 5, 5, 5, 5),
                   attr = c(1000, 1250, NA, 900, 1005, 1000,
                   1250, 950, 900, 1005, NA, 1250, 950, 900,
                   1005, 1000, 1250, NA, 900),
                   isUsed = c(T, T, T, T, T, T, T, T, T, T,
                   T, T, T, T, F, F, F, F, F))
}
}
\author{
Karin Wolken
}
