% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_cluster.R
\name{summarize_all_cluster}
\alias{summarize_all_cluster}
\title{Summarize all cluster sample}
\usage{
summarize_all_cluster(data, attribute = NA, element = TRUE, 
                             plotTot = NA, 
                             desiredConfidence = 0.95,
                             bernoulli = F)
}
\arguments{
\item{data}{data frame containing observations of variable of
interest for either cluster-level or plot-level data.}

\item{attribute}{character name of attribute to be summarized.}

\item{element}{logical true if parameter data is plot-level, false if
parameter data is cluster-level. Default is True.}

\item{plotTot}{numeric population size. Equivalent to the 
total number of possible elements in the population.}

\item{desiredConfidence}{numeric desired confidence level (e.g. 0.9).}

\item{bernoulli}{logical TRUE if data fitting the Bernoulli 
distribution is used.}
}
\value{
data frame of stand-level statistics including
standard error and confidence interval limits.
}
\description{
Summarizes population-level statistics for
cluster sample data. This function has two options: (1) 
Cluster sample with a normal distribution and (2) Cluster 
sample with a Bernoulli distribution.
}
\examples{
\dontrun{

# See Forest Sampling vignette for more details

# Plot level data can be expressed as:

plotLevelDataExample <- data.frame(clusterID = c(1, 1, 1, 1, 1, 2,  
                                                 2, 3, 4, 4, 4, 4, 
                                                 4, 4, 5, 5, 5, 5, 
                                                 5), 
                                   attr = c(1000, 1250, 950, 900, 
                                            1005, 1000, 1250, 950,
                                            900, 1005, 1000, 1250, 
                                            950, 900, 1005, 1000, 
                                            1250, 950, 900), 
                                   isUsed = c(T, T, T, T, T, T, T, 
                                              T, T, T, T, T, T, T, 
                                              F, F, F, F, F))

# Cluster level data can be expressed as: 

clusterLevelDataExample <- data.frame(clusterID = c(1, 2, 3, 4, 5), 
                                      clusterElements = c(4, 2, 9, 
                                                          4, 10), 
                                      sumAttr = c(1000, 1250, 950,
                                                  900, 1005), 
                                      isUsed = c(T, T, F, T, T))
# Set element = FALSE


# Bernoulli data can be expressed as: 

bernoulliData <- data.frame(plots = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
                            propAlive = c(0.75, 0.80, 0.80, 0.85, 
                                          0.70, 0.90, 0.70, 0.75, 
                                          0.80, 0.65))
# Set parameter bernoulli = TRUE

}
}
\author{
Karin Wolken
}
