% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleRandom.R, R/simple_random.R
\name{summarize_simple_random}
\alias{summarize_simple_random}
\alias{summarize_simple_random}
\title{Summarize simple random sample}
\usage{
summarize_simple_random(trainingData, attribute, popN,
  desiredConfidence = 0.9, infReplacement)

summarize_simple_random(trainingData, attribute, popN,
  desiredConfidence = 0.9, infReplacement)
}
\arguments{
\item{trainingData}{dataframe containing observations of variable of
interest, and stratum assignment for each plot}

\item{attribute}{character name of attribute to be summarized}

\item{popN}{numeric population size.}

\item{desiredConfidence}{numeric desired confidence level (e.g. 0.9)}

\item{infReplacement}{logical true if sample was done with replacement
or from an infite population. False if sampled without replacement,
from a finite population.}

\item{stratumTab}{dataframe containing acreages for each stratum}

\item{post}{logical true if post-stratification was used}

\item{trainingData}{dataframe containing observations of variable of
interest.}

\item{attribute}{character name of attribute to be summarized. Attribute
must already be expanded.}

\item{desiredConfidence}{numeric desired confidence level (e.g. 0.9).}
}
\description{
Summarizes population-level statistics for
simple random sample data. The calculations are derived from Chapter 3 in
Avery and Burkhart's (1967) Forest Measurements, Fifth Edition. The
variance terms refer to the variance of the mean, hence the
\code{n} terms in the deonminators.

Summarizes population-level statistics for
simple random sample data. The calculations are derived from Chapter 3 in
Avery and Burkhart's (1967) Forest Measurements, Fifth Edition. The
variance terms refer to the variance of the mean, hence the
\code{n} terms in the deonminators.
}
\examples{
\dontrun{
trainingData <- data.frame(bapa = c(120, 140, 160, 110, 100, 90),
  stratum = c(1, 1, 1, 2, 2, 2))
stratumTab <- data.frame(stratum = c(1, 2), acres = c(200, 50))
attribute = 'bapa'
desiredConfidence = 0.9
}
\dontrun{
trainingData <- data.frame(bapa = c(120, 140, 160, 110, 100, 90),
  plots = c(1, 2, 3, 4, 5, 6))
attribute = 'bapa'
desiredConfidence = 0.9
}
}
\author{
Karin Wolken

Karin Wolken
}
