% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_random.R
\name{summarize_simple_random}
\alias{summarize_simple_random}
\title{Summarize simple random sample}
\usage{
summarize_simple_random(data, attribute = "attr", popSize = NA,
  desiredConfidence = 0.95, infiniteReplacement = F)
}
\arguments{
\item{data}{data frame or vector containing observations of
variable of interest. Variable of interest must already be expanded
to the level of interest (e.g. stand-level).}

\item{attribute}{character name of attribute to be summarized.
Must be defined if data is input as a data frame.}

\item{popSize}{numeric population size. Defaults to NA (unknown popSize).}

\item{desiredConfidence}{numeric desired confidence level (e.g. 0.9).}

\item{infiniteReplacement}{logical true if sample was done with replacement
or from an infinite population. False if sampled without replacement,
from a finite population. Defaults to False.}
}
\value{
a data frame of population mean, variance, standard error, and
high and low confidence limits.
}
\description{
Summarizes population-level statistics for
simple random sample data. The calculations are derived from Chapter 3 in
Avery and Burkhart's (1967) Forest Measurements, Fifth Edition. The
variance terms refer to the variance of the mean, hence the
\code{sampleSize} terms in the denominators.
}
\examples{
\dontrun{
#trainingDataFrame
data <- data.frame(bapa = c(120, 140, 160, 110, 100, 90),
  plots = c(1, 2, 3, 4, 5, 6))
#trainingVector
data <- c(120, 140, 160, 110, 100, 90)
attribute <- 'bapa'
desiredConfidence <- 0.9
}
}
\author{
Karin Wolken
}
