% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic.R
\name{summarize_systematic}
\alias{summarize_systematic}
\title{Summarize systematic sample}
\usage{
summarize_systematic(trainingData, attribute = "attr", type = "vector",
  popSize = NA, desiredConfidence = 0.9)
}
\arguments{
\item{trainingData}{dataframe or vector containing observations of
variable of interest. Variable of interest must already be expanded
to the level of interest (e.g. stand-level).}

\item{attribute}{character name of attribute to be summarized.
Must be defined if data is input as a dataframe.}

\item{popSize}{numeric population size. Defaults to NA (unknown popSize).}

\item{desiredConfidence}{numeric desired confidence level (e.g. 0.9).}
}
\value{
a dataframe of population mean, variance, standard error, and
high and low confidence limits.
}
\description{
Summarizes population-level statistics for
systematic sample data. The calculations are derived from Chapter 3 in
Avery and Burkhart's (1967) Forest Measurements, Fifth Edition. The
variance terms refer to the variance of the mean, hence the
\code{sampleSize} terms in the denominators.
}
\examples{
\dontrun{
trainingData <- data.frame(bapa = c(120, 140, 160, 110, 100, 90),
  plots = c(1, 2, 3, 4, 5, 6))
attribute = 'bapa'
desiredConfidence = 0.9
}
}
\author{
Karin Wolken
}
