% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic.R
\name{summarize_systematic}
\alias{summarize_systematic}
\title{Summarize systematic sample}
\usage{
summarize_systematic(trainingData, attribute, popN, desiredConfidence = 0.9)
}
\arguments{
\item{trainingData}{dataframe containing observations of variable of
interest #not exactly necessary}

\item{desiredConfidence}{numeric desired confidence level (e.g. 0.9)}

\item{y}{attribute: vector of attribute to be summarized}
}
\description{
Summarizes population-level statistics for
systematic sample data. The calculations are derived from Chapter 3 in
Avery and Burkhart's (1967) Forest Measurements, Fifth Edition. The
variance terms refer to the variance of the mean, hence the
\code{n} terms in the deonminators.
}
\examples{
\dontrun{
trainingData <- data.frame(bapa = c(120, 140, 160, 110, 100, 90),
  plots = c(1, 2, 3, 4, 5, 6))
attribute = 'bapa'
desiredConfidence = 0.9
}
}
\author{
Karin Wolken
}
