% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3P.R
\name{summarize_threeP}
\alias{summarize_threeP}
\title{Summarize 3P sample}
\usage{
summarize_threeP(data, sampSize, treeCount, BAF, cvPercent, trueNetVBAR,
  height = FALSE, desiredConfidence = 0.95)
}
\arguments{
\item{data}{dataframe containing plot number, VBARs, and 
treeHeights}

\item{sampSize}{numeric}

\item{treeCount}{numeric}

\item{BAF}{numeric basal area factor used for sampling.}

\item{cvPercent}{numeric average Coefficient of Vairation 
expressed as a percent (e.g. 50)}

\item{trueNetVBAR}{numeric measured net VBAR (volume to basal
area ratio)}

\item{height}{boolean indicates that instead of the 
data containing treeHeight, it contains estimateVBAR}

\item{desiredConfidence}{numeric desired confidence level 
(e.g. 0.9).}
}
\value{
dataframe of statistics including standard error and 
confidence interval limits.
}
\description{
Summarizes population-level statistics for
3P sample data. The calculations are derived from Kim Iles
'A Sampler of Inventory Topics', pg 601.
}
\examples{
\dontrun{
data <- data.frame(plotNum = c(5, 4, 3, 5, 4),
                   VBARS = c(9, 8, 7, 8, 2),
                   treeHeight = c(1, 2, 3, 4, 5))
}
}
\author{
Karin Wolken
}
