% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage.R
\name{summarize_two_stage}
\alias{summarize_two_stage}
\title{Summarize two-stage sample}
\usage{
summarize_two_stage(data, plot = TRUE, attribute = NA,
  populationClusters = 0, populationElementsPerCluster = 0,
  desiredConfidence = 0.95)
}
\arguments{
\item{data}{dataframe containing observations of variable of
interest for either cluster-level of plot-level data.}

\item{plot}{logical True if parameter data is plot-level, False if
parameter data is cluster-level. Default is True.}

\item{attribute}{character name of attribute to be summarized.}

\item{desiredConfidence}{numeric desired confidence level (e.g. 0.9).}
}
\value{
dataframe of stand-level statistics including
standard error and confidence interval limits. All final values are
on a 'per plot' basis.
}
\description{
Summarizes population-level statistics for
two-stage sample data. The calculations are derived from Chapter 3 in
Avery and Burkhart's (1967) Forest Measurements, Fifth Edition. The
variance terms refer to the variance of the mean.
}
\examples{
\dontrun{
dataCluster <- data.frame(clusterID = c(1, 2, 3, 4, 5),
                          totClusterElements = c(5, 2, 1, 6, 5),
                          sampledElements = c(2, 2, 2, 2, 2),
                          isUsed = c(T, T, T, T, F),
                          attrSumCluster = c(1000, 1250, 950, 900, 1005))
summarize_two_stage(dataCluster, F, 'attr')

# realistic data
dataPlot <- data.framedata.frame(clusterID = c(1, 1, 1, 2, 2, 2, 3, 3, 3,  
                                           4, 4, 4, 5, 5, 5, 6, 6, 6), 
                             volume = c(500, 650, 610, 490, 475, 505, 
                                        940, 825, 915, 210, 185, 170, 
                                        450, 300, 500, 960, 975, 890),
                             isUsed = c(T, T, T, T, T, T, T, T, T, T, 
                                        T, T, T, T, T, T, T, T))
summarize_two_stage(redData, T, 'volume', populationClusters = 16,
                    populationElementsPerCluster = 160)

}
}
\author{
Karin Wolken
}
